[IMJ] Started working on 'ImageWrangler' which is going to be IMJ's regular image loading facility.

+++ IMJ/src/imj/LinearStorage.java
@@ -5,6 +5,7 @@ import static java.lang.Math.max;
 import static net.sourceforge.aprog.tools.Tools.unchecked;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
@@ -27,12 +28,32 @@ public final class LinearStorage extends Image.Abstract {
 	
 	private final FloatBuffer chunkFloats;
 	
+	private LinearStorage(final int rowCount, final int columnCount, final File file, final RandomAccessFile data) {
+		super(rowCount, columnCount);
+		
+		try {
+			this.file = file;
+			this.data = data;
+			this.chunkBytes = this.data.getChannel().map(MapMode.READ_WRITE, 0L, (HEADER_DATUM_COUNT + rowCount * columnCount) * DATUM_SIZE);
+			this.chunkInts = this.chunkBytes.asIntBuffer();
+			this.chunkFloats = this.chunkBytes.asFloatBuffer();
+		} catch (final Exception exception) {
+			throw unchecked(exception);
+		}
+	}
+	
 	public LinearStorage(final int rowCount, final int columnCount) {
+		this(rowCount, columnCount, true);
+	}
+	
+	public LinearStorage(final int rowCount, final int columnCount, final boolean deleteFileOnExit) {
 		super(rowCount, columnCount);
 		
 		try {
 			this.file = File.createTempFile("tiling", ".raw");
-			this.file.deleteOnExit();
+			if (deleteFileOnExit) {
+				this.file.deleteOnExit();
+			}
 			this.data = new RandomAccessFile(this.file, "rw");
 			this.chunkBytes = this.data.getChannel().map(MapMode.READ_WRITE, 0L, (HEADER_DATUM_COUNT + rowCount * columnCount) * DATUM_SIZE);
 			this.chunkInts = this.chunkBytes.asIntBuffer();
@@ -43,6 +64,10 @@ public final class LinearStorage extends Image.Abstract {
 		} catch (final Exception exception) {
 			throw unchecked(exception);
 		}
+	}
+	
+	public final File getFile() {
+		return this.file;
 	}
 	
 	public final void close() {
@@ -94,5 +119,17 @@ public final class LinearStorage extends Image.Abstract {
 	public static final int DATUM_SIZE = max(Integer.SIZE, Float.SIZE) / 8;
 	
 	public static final int HEADER_DATUM_COUNT = 2;
+	
+	public static final LinearStorage open(final File file) {
+		try {
+			final RandomAccessFile data = new RandomAccessFile(file, "rw");
+			final int rowCount = data.readInt();
+			final int columnCount = data.readInt();
+			
+			return new LinearStorage(rowCount, columnCount, file, data);
+		} catch (final IOException exception) {
+			throw unchecked(exception);
+		}
+	}
 	
 }
